+++ src/main/java/fuzs/deathfinder/config/ServerConfig.java
@@ -38,6 +38,9 @@ public class ServerConfig extends AbstractConfig {
         public boolean distanceComponent = true;
         @Config(description = "Who should be allowed to click the position component to teleport there.")
         public TeleportRestriction allowTeleporting = TeleportRestriction.EVERYONE;
+        @Config(description = {"Amount of seconds in which teleporting to the last death point is possible.", "Set to -1 to remove time limit."})
+        @Config.IntRange(min = -1)
+        public int teleportInterval = 300;
 
         public ComponentsConfig() {
             super("death_message_components");