+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v3.1.0-1.18.1] - 2022-02-20
+### Changed
+- All players can now teleport to their death point, not just operators (this can be changed back in the config)
+- To avoid abuse of this feature, a time limit can be set for how long after dying this is possible, also only the latest death point can be teleported to
+
 ## [v3.0.0-1.18.1] - 2021-12-13
 - Ported to Minecraft 1.18
 ### Removed